Refactor DataSource so that differetn device types can override features.
Move creation to a factory and make it part ofthe DaemonInterface so
that it can be changed when the device type changes.

Add implementation for generic huami and zeppos devices
