+++ ui/src/dataproviderfactory.h
@@ -1,11 +1,13 @@
 #ifndef DATAPROVIDERFACTORY_H
 #define DATAPROVIDERFACTORY_H
 
+#include "datasource.h"
 
 class DataProviderFactory
 {
 public:
-    DataProviderFactory();
+    static DataSource* dataSource(const QString &deviceType);
+
 };
 
 #endif // DATAPROVIDERFACTORY_H
