+++ ui/src/zepposdatasource.cpp
@@ -1,0 +1,12 @@
+#include "zepposdatasource.h"
+
+
+ZeppOSDataSource::ZeppOSDataSource()
+{
+    qDebug() << Q_FUNC_INFO;
+}
+
+QList<DataSource::SleepSession> ZeppOSDataSource::calculateSleep(const QDate &day)
+{
+
+}
